move cross chain choose to staked.cpp
