+++ src/crosschain.h
@@ -13,8 +13,6 @@ typedef struct CrosschainAuthority {
     size_t requiredSigs;
 } CrosschainAuthority;
 
-CrosschainAuthority auth_STAKED_chosen(const char *notaries_chosen[][2],int num_notaries);
-
 int GetSymbolAuthority(const char* symbol);
 bool CheckTxAuthority(const CTransaction &tx, CrosschainAuthority auth);
