+++ src/notarisationdb.cpp
@@ -51,13 +51,6 @@ NotarisationsInBlock ScanBlockNotarisations(const CBlock &block, int nHeight)
             {
               // era 1
               auth_STAKED = auth_STAKED_chosen(notaries_STAKED1,num_notaries_STAKED1);
-              /*
-              auth_STAKED.requiredSigs = (num_notaries_STAKED1 / 5);
-              auth_STAKED.size = num_notaries_STAKED1;
-              for (int n=0; n<auth_STAKED.size; n++)
-                  for (size_t i=0; i<33; i++)
-                      sscanf(notaries_STAKED1[n][1]+(i*2), "%2hhx", auth_STAKED.notaries[n]+i);
-              */
             } else if (staked_era == 2)
             {
               // era 2
